requirements.txt
astroid==2.3.3
dominantcolors==0.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.2
Pillow==7.0.0
pkg-resources==0.0.0
psutil==5.7.0
pycairo==1.18.2
PyGObject==3.34.0
pylint==2.4.4
PyYAML==5.3.1
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
xdg==4.0.1
